pipeline/requirements.txt
APScheduler==3.10.4
datasketch==1.6.4
confluent-kafka==2.3.0
pyyaml==6.0.1
requests==2.31.0
web3==6.15.1
pycardano==0.11.0
